fixed movegen bugs, added move_isLegal, moves are now filtered

+++ src/move.h
@@ -25,4 +25,4 @@ void move_unmake(BoardState* state, Move* m);
 int8_t move_gen(BoardState* state, Move* list);
 int8_t move_genPiece(BoardState* state, Move* list, int8_t sq, int8_t val);
 
-int move_isLegal(BoardState* state, Move* move);
+int8_t move_isLegal(BoardState* state, Move* move);
